Move dto objects to models package and Session to firebase
