+++ app/src/androidTest/java/com/example/musicmap/feed/MusicMemoryTest.java
@@ -6,6 +6,8 @@ import android.net.Uri;
 
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
+import com.example.musicmap.models.MusicMemory;
+import com.example.musicmap.models.Song;
 import com.google.firebase.firestore.GeoPoint;
 
 import org.junit.Before;
