+++ app/src/main/java/com/example/musicmap/models/Artist.java
@@ -1,4 +1,4 @@
-package com.example.musicmap.user;
+package com.example.musicmap.models;
 
 import androidx.annotation.NonNull;
