+++ app/src/main/java/com/example/musicmap/screens/main/FeedFragment.java
@@ -11,7 +11,7 @@ import android.widget.ListView;
 
 import com.example.musicmap.R;
 import com.example.musicmap.util.adapters.FeedAdapter;
-import com.example.musicmap.feed.MusicMemory;
+import com.example.musicmap.models.MusicMemory;
 import com.example.musicmap.util.firebase.Queries;
 
 import java.util.ArrayList;
