+++ app/src/main/java/com/example/musicmap/screens/main/HomeActivity.java
@@ -12,8 +12,8 @@ import com.example.musicmap.R;
 import com.example.musicmap.screens.artist.ArtistDataFragment;
 import com.example.musicmap.screens.map.PostMapFragment;
 import com.example.musicmap.screens.profile.ProfileActivity;
-import com.example.musicmap.user.Session;
-import com.example.musicmap.user.User;
+import com.example.musicmap.firebase.Session;
+import com.example.musicmap.models.User;
 import com.example.musicmap.util.Constants;
 import com.example.musicmap.util.permissions.LocationPermission;
 import com.example.musicmap.util.spotify.SpotifyAuthActivity;
