+++ app/src/main/java/com/example/musicmap/screens/map/ArtistDataMapFragment.java
@@ -2,9 +2,9 @@ package com.example.musicmap.screens.map;
 
 import android.util.Log;
 
-import com.example.musicmap.user.Artist;
-import com.example.musicmap.user.Session;
-import com.example.musicmap.user.User;
+import com.example.musicmap.models.Artist;
+import com.example.musicmap.firebase.Session;
+import com.example.musicmap.models.User;
 import com.example.musicmap.util.firebase.Queries;
 import com.example.musicmap.util.map.MusicMemoryOverlay;
