+++ app/src/main/java/com/example/musicmap/screens/profile/ProfileActivity.java
@@ -9,7 +9,7 @@ import com.example.musicmap.R;
 import com.example.musicmap.SessionAndInternetListenerActivity;
 import com.example.musicmap.screens.main.HomeActivity;
 import com.example.musicmap.screens.settings.SettingsActivity;
-import com.example.musicmap.user.Session;
+import com.example.musicmap.firebase.Session;
 import com.example.musicmap.util.Constants;
 import com.example.musicmap.util.ui.FragmentUtil;
