+++ app/src/main/java/com/example/musicmap/screens/profile/ProfilePageFragment.java
@@ -16,9 +16,9 @@ import androidx.fragment.app.Fragment;
 import com.example.musicmap.R;
 import com.example.musicmap.R.id;
 import com.example.musicmap.util.adapters.FeedAdapter;
-import com.example.musicmap.feed.MusicMemory;
-import com.example.musicmap.user.Session;
-import com.example.musicmap.user.User;
+import com.example.musicmap.models.MusicMemory;
+import com.example.musicmap.firebase.Session;
+import com.example.musicmap.models.User;
 import com.example.musicmap.util.Constants;
 import com.example.musicmap.util.firebase.AuthSystem;
 import com.example.musicmap.util.firebase.Queries;
