+++ app/src/main/java/com/example/musicmap/util/adapters/FeedAdapter.java
@@ -14,10 +14,10 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import com.example.musicmap.R;
-import com.example.musicmap.feed.MusicMemory;
+import com.example.musicmap.models.MusicMemory;
 import com.example.musicmap.screens.main.MusicMemoryActivity;
 import com.example.musicmap.screens.profile.ProfileActivity;
-import com.example.musicmap.user.User;
+import com.example.musicmap.models.User;
 import com.example.musicmap.util.Constants;
 import com.example.musicmap.util.firebase.AuthSystem;
 import com.example.musicmap.util.ui.CircleTransform;
