+++ app/src/main/java/com/example/musicmap/util/adapters/PopularSongsAdapter.java
@@ -12,7 +12,7 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import com.example.musicmap.R;
-import com.example.musicmap.feed.Song;
+import com.example.musicmap.models.Song;
 import com.example.musicmap.screens.artist.SongCount;
 import com.squareup.picasso.Picasso;
