+++ app/src/main/java/com/example/musicmap/util/firebase/Queries.java
@@ -1,8 +1,8 @@
 package com.example.musicmap.util.firebase;
 
-import com.example.musicmap.feed.MusicMemory;
-import com.example.musicmap.feed.Post;
-import com.example.musicmap.feed.Song;
+import com.example.musicmap.models.MusicMemory;
+import com.example.musicmap.models.Post;
+import com.example.musicmap.models.Song;
 import com.example.musicmap.screens.artist.SongCount;
 import com.google.android.gms.tasks.Task;
 import com.google.android.gms.tasks.TaskCompletionSource;
